# fix: Improve performance of get_group_results query