--- a/.gitignore
+++ b/.gitignore
@@ -8,9 +8,14 @@ __pycache__
 .ruff_cache/
 .build/
 .venv/
+venv/
+.scratch/
 
 # vi
 *.swp
+
+# Database
+treeherder_backup_*
 
 # Caches
 .eslintcache