--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -67,7 +67,7 @@ services:
     volumes:
       - .:/app
     ports:
-      - '5000:5000'
+      - '${FRONTEND_PORT:-5000}:5000'
     platform: linux/amd64
 
   postgres:
@@ -78,10 +78,12 @@ services:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=mozilla1234
       - POSTGRES_DB=treeherder
+    labels:
+      - prune=False
     volumes:
       - postgres_data:/var/lib/postgresql/data
     ports:
-      - '5432:5432'
+      - '${POSTGRES_PORT:-5432}:5432'
 
   redis:
     container_name: redis
@@ -92,7 +94,7 @@ services:
     # Hide Redis `notice` log level startup output spam.
     command: redis-server --loglevel warning
     ports:
-      - '6379:6379'
+      - '${REDIS_PORT:-6379}:6379'
 
   rabbitmq:
     container_name: rabbitmq