--- a/pyproject.toml
+++ b/pyproject.toml
@@ -31,6 +31,10 @@ target-version = "py310"
 # In addition to the standard set of exclusions, omit all tests, plus a specific file.
 extend-exclude = ["*/.*/",".*/","__pycache__","node_modules"]
 
+# Also lint/format pyi files
+extend-include = ["*.pyi"]
+
+[tool.ruff.lint]
 select = [
   # pycodestyle
   "E",
@@ -48,9 +52,6 @@ ignore = [
   "E501",
 ]
 
-# Also lint/format pyi files
-extend-include = ["*.pyi"]
-
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 # Ignore `module-import-not-at-top-of-file` rule of `pycodestyle`
 "treeherder/model/models.py" = ["E402"]