--- a/tests/log_parser/test_store_failure_lines.py
+++ b/tests/log_parser/test_store_failure_lines.py
@@ -199,7 +199,7 @@ def test_group_status_duration(activate_responses, test_repository, test_job):
     assert log_obj.groups.count() == 29
 
 
-def test_get_group_results(activate_responses, test_repository, test_job):
+def test_get_group_results(activate_responses, test_job):
     log_path = SampleData().get_log_path("mochitest-browser-chrome_errorsummary.log")
     log_url = "http://my-log.mozilla.org"
 
@@ -209,13 +209,13 @@ def test_get_group_results(activate_responses, test_repository, test_job):
     log_obj = JobLog.objects.create(job=test_job, name="errorsummary_json", url=log_url)
     store_failure_lines(log_obj)
 
-    groups = get_group_results(test_repository, test_job.push)
+    groups = get_group_results(test_job.repository, test_job.push)
     task_groups = groups["V3SVuxO8TFy37En_6HcXLs"]
 
     assert task_groups["dom/base/test/browser.ini"]
 
 
-def test_get_group_results_with_colon(activate_responses, test_repository, test_job):
+def test_get_group_results_with_colon(activate_responses, test_job):
     log_path = SampleData().get_log_path("xpcshell-errorsummary-with-colon.log")
     log_url = "http://my-log.mozilla.org"
 
@@ -225,7 +225,7 @@ def test_get_group_results_with_colon(activate_responses, test_repository, test_
     log_obj = JobLog.objects.create(job=test_job, name="errorsummary_json", url=log_url)
     store_failure_lines(log_obj)
 
-    groups = get_group_results(test_repository, test_job.push)
+    groups = get_group_results(test_job.repository, test_job.push)
     task_groups = groups["V3SVuxO8TFy37En_6HcXLs"]
 
     assert task_groups[
@@ -259,7 +259,7 @@ def create_errorsummary_job(base_job, create_jobs, log_filenames):
 
     job_defs = []
     urls = []
-    for log_filename in log_filenames:
+    for idx, log_filename in enumerate(log_filenames):
         log_path = SampleData().get_log_path(log_filename)
         log_url = f"http://my-log.mozilla.org/{log_path}"
 
@@ -277,20 +277,16 @@ def create_errorsummary_job(base_job, create_jobs, log_filenames):
                 "status": "completed",
                 "result": "success" if "_pass" in log_filename else "testfailed",
                 "name": f"{job_def['job']['name']}{task_ending}",
-                "reference_data_name": job_def["job"]["reference_data_name"].replace(
-                    "a", str(random.randint(0, 9))
-                ),
+                "reference_data_name": job_def["job"]["reference_data_name"].replace("a", str(idx)),
                 "job_guid": job_def["job"]["job_guid"]
-                .replace("e", str(random.randint(0, 9)))
-                .replace("d", str(random.randint(0, 9))),
+                .replace("e", str(idx))
+                .replace("d", str(idx * 10 + random.randint(0, 9))),
                 "start_timestamp": job_def["job"]["start_timestamp"]
                 + 100
-                + random.randint(0, 100)
-                + random.randint(0, 100),
-                "taskcluster_task_id": job_def["job"]["taskcluster_task_id"].replace(
-                    "T", str(random.randint(0, 9))
-                ),
-                "taskcluster_retry_id": "0",
+                + idx * 50
+                + random.randint(0, 25),
+                "taskcluster_task_id": job_def["job"]["taskcluster_task_id"].replace("T", str(idx)),
+                "taskcluster_retry_id": str(idx),
             }
         )
         job_defs.append(job_def)