@camd 2025-08-17
I did a test with the old query and the new query:

```
[2025-08-17 03:50:14,702] INFO [treeherder.webapp.api.push:508] get_group_results took 0.207 seconds for push 56fa7d27a8b4d50953b21af0e70f3d53bde8a82e

[2025-08-17 03:50:14,807] INFO [treeherder.webapp.api.push:515] get_group_results_new took 0.105 seconds for push 56fa7d27a8b4d50953b21af0e70f3d53bde8a82e
```

So that looked promising.  But then I reversed the order that I queried it (doing new first, and old second)

```
[2025-08-17 03:53:30,183] INFO [treeherder.webapp.api.push:509] get_group_results_new took 0.120 seconds for push 56fa7d27a8b4d50953b21af0e70f3d53bde8a82e

[2025-08-17 03:53:30,311] INFO [treeherder.webapp.api.push:516] get_group_results took 0.128 seconds for push 56fa7d27a8b4d50953b21af0e70f3d53bde8a82e
```
So there must be some kind of caching going on.  Will investigate further...