@camd 2025-08-17
I think my local DB dataset is just too small to get a valid timing measurement...

Working now on using either staging or prod DB to test.